Fixed schema not using correct data type for id's

+++ src/main/resources/db/migration/V001__initial_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-    user_id serial primary key,
+    user_id bigserial primary key,
     sub varchar(255) not null,
     username varchar(255) not null,
     picture varchar(255),
@@ -10,8 +10,8 @@ CREATE TABLE users (
 );
 
 CREATE TABLE servers (
-    server_id serial primary key,
-    owner_id int REFERENCES users(user_id),
+    server_id bigserial primary key,
+    owner_id bigint REFERENCES users(user_id),
     name varchar(255) not null,
     description varchar(255),
     created_at timestamp not null default now(),
@@ -19,8 +19,8 @@ CREATE TABLE servers (
 );
 
 CREATE TABLE channels (
-    channel_id serial primary key,
-    server_id int not null REFERENCES servers(server_id),
+    channel_id bigserial primary key,
+    server_id bigint not null REFERENCES servers(server_id),
     name varchar(255) not null,
     description varchar(255),
     created_at timestamp not null default now(),
@@ -28,17 +28,17 @@ CREATE TABLE channels (
 );
 
 CREATE TABLE messages (
-    message_id serial primary key,
-    channel_id int not null REFERENCES channels(channel_id),
-    user_id int not null REFERENCES users(user_id),
+    message_id bigserial primary key,
+    channel_id bigint not null REFERENCES channels(channel_id),
+    user_id bigint not null REFERENCES users(user_id),
     content varchar(255) not null,
     created_at timestamp not null default now(),
     updated_at timestamp not null default now()
 );
 
 CREATE TABLE server_user (
-    server_id int not null REFERENCES servers(server_id),
-    user_id int not null REFERENCES users(user_id),
+    server_id bigint not null REFERENCES servers(server_id),
+    user_id bigint not null REFERENCES users(user_id),
     primary key (server_id, user_id),
     created_at timestamp not null default now()
 );
